+++ src/Regular.cpp
@@ -10,7 +10,7 @@ Regular::Regular(int numSamples)
 
 void Regular::generateSamples()
 {
-	const int n = static_cast<int>(sqrtf(numSamples_));
+	const int n = static_cast<int>(sqrtf(static_cast<float>(numSamples_)));
 
 	for (int i = 0; i < numSets_; i++)
 		for (int p = 0; p < n; p++)
